requirements.txt
python-dotenv
python-slugify
requests
selenium